+++ src/pages/UsersPage/pages/SingleUserPage/components/UserDetails/components/Posts/components/Post/Post.module.scss
