+++ src/pages/UsersPage/pages/SingleUserPage/components/UserDetails/components/Posts/components/Post/hooks/useDeletePost.ts
