+++ src/pages/UsersPage/pages/SingleUserPage/components/UserDetails/components/Posts/components/index.ts
