+++ src/pages/UsersPage/pages/SingleUserPage/components/UserDetails/components/Posts/types/index.ts
