+++ src/pages/UsersPage/pages/SingleUserPage/components/UserDetails/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Posts } from "./Posts/Posts";
